Refactor suggestion services folder structure
