+++ cmd/suggestion/nas/enas/v1alpha3/main.py
@@ -4,7 +4,7 @@ import time
 
 from pkg.apis.manager.v1alpha3.python import api_pb2_grpc
 from pkg.apis.manager.health.python import health_pb2_grpc
-from pkg.suggestion.v1alpha3.nas.enas_service import EnasService
+from pkg.suggestion.v1alpha3.nas.enas.enas_service import EnasService
 
 
 _ONE_DAY_IN_SECONDS = 60 * 60 * 24
