+++ test/suggestion/v1alpha3/test_skopt_service.py
@@ -5,7 +5,7 @@ import unittest
 from pkg.apis.manager.v1alpha3.python import api_pb2_grpc
 from pkg.apis.manager.v1alpha3.python import api_pb2
 
-from pkg.suggestion.v1alpha3.skopt_service import SkoptService
+from pkg.suggestion.v1alpha3.skopt.skopt_service import SkoptService
 
 
 class TestSkopt(unittest.TestCase):
